refactor(session): replace cookie secret storage with localStorage
Instead of using cookies to store the secret with which the session data
is encrypted (and relying on their automatic expiration to expire the
session), store the secret in localStorage instead and manually handle
the expiration.

The main reason for this change is that when using a solution such as
Cordova to create 'native' apps, cookies don't exist. LocalStorage on
the other hand is available in these cases as well and without
localStorage storing sessions would not work anyway.